Switch to Next.js
This commit updates the site to render via Next.js rather than Hugo.
After working with Hugo for a few years on the Tailscale website, I'm
a bit sick of it and its opininated nature. It's great for a standard
blog setup, but it's hard to work around it's quirks if you want to
do something else.

I'm switching my site to Next.js for (a) the practice and (b) the
extensibility it provides for introducing new sections of content.