Add MDX to posts
This commit adds MDX support to the site. It took me some time to
warm up to MDX, but I see the benefits.

Not only does it simplify shortcode-like blocks in a more standard form,
it also enables some nice optimizations like Nextjs' image handling
features. It also lays the groundwork for me to use interactive
components in posts without a lot of manual effort.