Upgrade to Next.js 13
This commit upgrades to the latest version of Next.js, including
removing <a> tag wrappers around <Link> tags and removing flags for
experimental features that are now supported.

This commit doesn't upgrade to the app directory / react server
components, since there's a long tail of issues with contentlayer and
rsc still to be fixed.