Migrate to Next.js app directory (#25)
This allows us to use new the server components feature, which means
less JS gets sent to clients and more work is done on the server-side.
A nice, easy win.

It's not a perfect migration, since Next.js app directory still has a
few rough edges. For instance, we have a `[...slug]` page that acts as a
polyfill for a proper catch-all 404 page. And next/dynamic imports aren't
supported yet, but putting elements in a suspense block seems to do
something similar.